Refactor folder names and imports
