+++ src/schema/organization/queries.js
@@ -1,14 +1,16 @@
 import { GraphQLString, GraphQLNonNull } from 'graphql';
 
-import OrganizationType from './OrganizationType';
+import OrganizationType from './types';
 import resolvers from './resolvers';
 
 const organization = {
   type: OrganizationType,
-  args: { login: { 
-    type: new GraphQLNonNull(GraphQLString),
-    description: 'Organization login on Github'
-  }},
+  args: {
+    login: {
+      type: new GraphQLNonNull(GraphQLString),
+      description: 'Organization login on Github'
+    }
+  },
   resolve: resolvers.Query.organization
 }
