+++ src/schema/organization/types.js
@@ -5,12 +5,14 @@ import {
   GraphQLInt,
 } from 'graphql';
 
-import members from './members';
-import repositories from './repositories';
-import teams from './teams'
+import members from './members/queries';
+import repositories from './repositories/queries';
+import teams from './teams/queries'
 
-export default new GraphQLObjectType({
+const OrganizationType = new GraphQLObjectType({
   name: 'Organization',
+  description: '',
+
   fields: () => ({
     id: { type: GraphQLID },
     name: { type: GraphQLString },
@@ -26,3 +28,5 @@ export default new GraphQLObjectType({
     repositories,
   })
 });
+
+export default OrganizationType;
